Additional changes for EXTERNAL dummy arguments
In lower_function() and handle_arguments(), add additional conditions
for using a procedure dummy's descriptor argument to access the
procedure's closure pointer. If the procedure dummy argument is
declared EXTERNAL, then do not use the descriptor argument by default
since it could be a non-Fortran procedure. If the procedure dummy
argument is not declared EXTERNAL or is an interface, then assume
it is a Fortran procedure and use the descriptor to access the
closure pointer.

XBIT(54, 0x20) is used to override this restriction and always
access the closure pointer from the procedure dummy's descriptor
argument. That is, assume that the procedure dummy is always a
Fortran routine when XBIT(54, 0x20) is set.

Add proc_arg_needs_proc_desc() that tests the criteria noted above
to symutl.c.