Fix undefined behaviour in scan.c:alpha()
The code initialised a pointer to a value that according to C is undefined
behavior (section 6.5.6, clause 8):

  char *ip = id - 1; /* point 1 before id[] */

This variable is then conditionally incremented and written to in a
do-while loop, but if the condition was never true, the value '\0' is
written to address 'id - 1'.

The functionality of alpha() breaks when compiled with LLVM 7.