Fix a recent regression in atomics support
A recent update changed the types of the bit-fields in ATOMIC_INFO from
unsigned int to enum types.  This resulted in the values of those
fields being incorrectly sign extended when extracted from the struct.
Fix the problem by changing the fields of ATOMIC_INFO to not be
bit-fields.  This required changing the code that encodes/decodes
ATOMIC_INFO to/from an int.