Add another guard for the previous check-in for resolveImp() function
Do not overload the symbol if its scope is the same as the current
scope. That is, in the case where symbol is an interface, its scope
is its enclosing module (when the interface is defined in a
module). If the symbol is a procedure, then its scope is itself.
Therefore, also look at SCOPEG(scope) in this case.