Fix use of immediate values in Intel SSE intrinsics
FMA3 version of double precision POWER().

Newer versions of the clang compiler require that immediate values
of Intel SSE intrinsics be correctly sized. That is, if an immediate
field is 8 bits wide, the unsigned constant must be <= 255.

For the double precision routines in pow/fma3 eliminate using the
constant H55_D and replace with explicit immediate constants. This
immediate value is used to index within vector registers; these
positions will ever change.