Resolve issue with OpenMP atomics
With have OMP FLUSH with OMP ATOMIC operations;
there is an ICE regarding dtype being zero.

For example, for the following piece of code:

       !$OMP FLUSH(this)
             if(.not.res) then
       !$OMP ATOMIC READ
              n=this%num_out_oprnds
       !$OMP ATOMIC READ
              m=this%num_accumulated

With OMP FLUSH(this) here, "this" is set to be VOLATILE.
Later, in expand, the compiler sets nmex to NME_VOL.
This causes dtype to be DT_NONE "exp_mp_atomic_read".
Relying on NME to get the dtype of atomic variables
does not work in cases like this. When this happens,
get the dtype info from the symbols from ILMs.