Fix seg fault related to procedure dummy arguments
In transform_call() of rest.c, make sure to use the procedure dummy
argument's dummy descriptor, not a temporary descriptor. If the
argument is a procedure/procedure pointer and not a dummy argument,
then use a temporary descriptor.
Also if calling an internal procedure, generate a temporary
descriptor for the call.

In is_proc_desc_arg() of exp_rte.c:

Add some additional code for extracting the symbol from ILTY_LOAD
and IL_ACON instructions.

In mk_argasym() of exputil.c, propagate the IS_PROC_DESC
flag from the dummy argument to the stack temporary variable.