Compilation with -Msecond_underscore has three types of failures:
 1) Code that uses symbols from ieee_arithmetic, iso_c_binding, and
    omp_lib modules, among others, which get resolved in support
    libraries, fails.  The problem is that the compiler adds double
    underscores to these names, although names in the libraries don't.
    This is the failure case noted in the fs.

 2) Code that passes intrinsic routines as arguments fails.  Again,
    the problem is that the compiler adds double underscores to these
    routine names, but library routines for these intrinsics don't.

 3) Code with finalizers fails.  The problem is that the compiler
    generates one set of finalizer linker names directly from
    corresponding symbols, but also generates instances of these
    same symbols from component pieces.  With double underscores,
    the two sets of symbols don't match.

The fix for #1 and #2 is to suppress double underscore suffixes for
library symbols using a new LIBSYM sym flag.  The code that sets
this flag uses string compares.  These string compares are done
under narrowly targetted conditions, including the presence of
-Msecond_underscore.  This should limit the risk of false positive
recognition of library symbol names in user name space to rare cases.

The fix for #3 is to change the code that processes finalizer names
in terms of components to match the results generated from normal
symbol names.