Improve handling of protected array defined in parent module
Regression occured when using a protected array defined in
the parent module in a submodule. The compiler has to check
all data types. In this fix, move the is_used_by_submod()
check inside the set_assn(). This change will enable checks
for all module protected members for the submodule.