Fix issue with type-bound procedure name resolution
When two different modules with type bound
procedures share the same interface name (func_iface)
are bound to an abstract type, and instances of
this abstract type are made and when a call to the
interface is made using this instance of abstract
type, the name of the procedure gets resolved to just
the abstract interface name, since it doesn't know
the actual type which extends the abstract type.

Along with that, the front end does not set
the inmodule value for the interface function
func_iface. This leads to the function name not
being mangled with the corresponding module name and
hence the symbol names clash in the LLVM bridge. This
clash in the symbol table between the two interface
function symbols causes the signature (abi) of the
first function to be overwritten by the second since
their names match (due to no name mangling). This
causes incorrect LLVM IR generation.

Setting of inmodule flag allows the compiler to
generate the inmodule sptr for func_iface interface
name in the ILM and help the LLVM bridge generate the
right mangled name for the interface name thereby
differentiating between the abstract interfaces
belonging to the two modules.

This fix passes a new bit IS_INTERFACE from the
front end to the back end to allow the LLVM bridge to
recognize these cases and behave as expected. Before,
the INMODULE flag would not be set before for abstract
interfaces (which led to other issues); this extra
information is needed.