Add lexical scopes for do concurrent loops
A problem occurs in code that compiles a do concurrent loop
as a parallel OpenMP loop with a collapse clause, like this:

    !$omp do collapse(2)
    do concurrent (i=1:m, j=1:n)
      a(i,j) = 1
    enddo

With the collapse clause, parallel code generation for this loop (nest)
is missing the upper scope label. This would likely cause problems for
as yet unimplemented parallel code generation. Without parallelization
but with -g the missing label causes an internal compiler error when
DWARF generation assumes that the scope labels are present and valid.
The fix for this problem is to propagate scoping information in routine
collapse_add. (semutil.c)

Std dumps are also updated to include lexical block syms. (dump.c)