c_associated calls with expression arguments
Compilation of the following call in the bug test case generates an
internal compiler error:

  c_associated(xc_func_info_get_ref(info%ptr, number))

The problem is that the argument to c_associated is a function call
rather than a simple identifier (which the compiler handles correctly).

Test nag_f08_o op1 contains the call:

  c_associated(tmp,c_loc(b(1)))

Because the second argument is not a simple identifier, the compiler
effectively ignores it.  In the test, tmp has the same non-null value
as c_loc(b(1)), so c_associated(tmp) is TRUE, and the test compiles and
executes "correctly".  The test fails if the two arguments are swapped.

The fix is to temporize expression arguments to c_associated, so that
arguments are always simple identifiers.