Removed an assertion that was not strictly accurate
The LLVM shufflevector returns a type with the number of
elements within the mask and the element type of the input
vectors, but the VPERMUTE ILI does not totally conform to
this model. For example, lhs_ili might be a compare operand
(VCMP) whose dtype represents the dtype of its comparators,
not the result type; and the result of a (vector) compare
in LLVM is always a bit-mask.