Fix nested c_loc function calls with expression arguments
There is an additional issue similar to the problem of c_associated
calls with expression arguments addressed by the prior change to file
semfunc.c. Compilation with inlining of the following call also
generates an internal compiler error:

  f(c_loc(x(5)))

The problem for this expression is that the compiler has special case
code for (by reference) c_loc argument expressions like this that
effectively removes the inner c_loc call, so the call looks like
f(x(5)). This is valid if the call is actually materialized as usual.
But if the outer call is inlined (routine f here), it is then invalid
to remove the c_loc intermediary for expressions such as x(5). The
fix again involves temporization to turn an expression into a simple
identifier. These two temporization cases are in some sense inverses.
The earlier case was temporization of expression arguments to intrinsic
c_associated. This latter case is temporization of an expression
argument to c_loc when c_loc is itself an argument to another call
(possibly c_associated).