Detect the special case for ISHFT for int8
When we convert int to signed short int, we first do a
sign extension, and then left shift 24 bits, followed by
arithmetic right shift 24 bits (filled with sign bits),
then followed by the real ishft, then truncate to
int8. Therefore the logical left shift never happened
to int8.

The fix is to recognize the special case for ISHFT on a
short int, and do a logical shift instead of an arithmetic
shift.