Fixes for SUBMODULE, MODULE and tests
Fix internal compiler error with new_symbol not found with submodules

If the compiler finds one type bound procedure is bound to
the separate module procedure, export the separate module
procedure dummy arguments. Otherwise, those symbols are
not visible for other modules to use.

INMODULE bit was also not set, which causes the separate
module procedure name is not mangled properly in the
backend that causes an undefined error. This only happens
when the SMP is declared and defined within the same module.

For ST_USERGENERIC symbols, do not use the ST_ALIAS symbol
table type for the symbol if its GTYPE field is set. GTYPE
is set when the ST_USERGENERIC name overloads a derived
type name.