Fix a regression in debug metadata
Fix an issue which blocked the local binding that
connects the !DILocalVariable to @llvm.dbg.declare. It
was introduced in 7da0254f387 which intended to eliminate
spurious local variables in the DWARF. This fix keeps
the compiler created variable's name unexposed to the
DWARF and restores the local binding as well.