Fix #609 issue with variable addresses in internal subroutines
An adjustable character array is not recognized within a
subroutine nested inside this parent subroutine.

The INTERNREF flag was not being set for the char array
sptr. This is because the logic which sets INTERNREF
among other things only checked for arrays, pointers,
allocatables, proc dummy but not adjustable char
arrays. Checking for 'ADJLEN' bit allows the compiler
to set INTERNREF bit flag.  The len sptr also needs
the INTERNREF flag to be set.