Fix for segfault in multiple entries OpenMP functions
This patch adds missing check for multiple entries in a function
that builds prologue for a Fortran function (or subroutine) that
contains at least one OpenMP directive.

Since ll_save_gtid_val() function usually used for single-entry
Fortran functions can handle only one entry, processing of multiple
entries is handled similarly in a for loop.

A short test case which validates this patch is also added
to the commit.

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>