Fix in preprocessing for Flexi app
The preprocessor in flang detects 1.NE.PP_NUM as a number.
The fix uses the knowledge that a constant number can have
atmost one dot character in it. If more than one dot character
is seen then the scanning has gone too far ahead.
This fix removes PP_NUM from the number detected and allows
for substitution of PP_NUM with the defined preprocessor
value.