Transform ubound, lbound calls to value references
For fortran adjustable arrays with known expression
as the bounds, it should be possible to replace the
bounds call with the actual expression.

For example:
  subroutine arr_bound(s1,s2,s3,n1,n2,n3,A)
  implicit none
  integer, intent(in) :: s1,s2,s3,n1,n2,n3
  real(8), intent(inout) :: A(s1:n1,s2:n2,s3:n3)
  integer :: lb1,ub1
  lb1 = lbound(A,1)
  ub1 = ubound(A,1)

In the case above, the dimensions of A are
(s1:n1,s2:n2,s3:n3). Thus a call to lbound(A,1) is
value s1. And call of ubound(A,1) is value n1. The
frontend should be capable of doing this simple
replacement.