Fix min/max operations to generate correct code
This cleans up the mappings from CC_RELATION codes to the LLVM
condition codes. This code was rewritten in a more principled
way than it had been previously. Specifically, for floating-point
operations such as "lt" and "not-ge" should not be considered the
same. "lt" should return false is either argument is a NaN.
"not-ge" would return true if either argument is a NaN.

For the bridge, min and max operators put the alternative in the
else part of the expression. That is,
  max(a,b) -->  max = a; if a >= b then nop else max = b
So, we only want the "max = b" part to execute when neither a
nor b is a NaN and a < b. In LLVM parlance, correctness requires
that the UGE comparison operator be used.

There are a few extra changes related to fixing some of the copious
warnings from this file, which were made while trying to beef up
the type correctness of the module.