Properly mark implicit firstprivate for OpenMP task construct
Fortran frontend does not mark variables with
firstprivate implicitly even though the OpenMP spec
says in P:182 L:14 "In a task generating construct,
if no default clause is present, a variable for which
the data-sharing attribute is not determined by the
rules above is firstprivate."

In the following code, P should be firstprivate
by default because it is private for the parallel
region. However, Flang compiler does not make P
implicitly firstprivate. It causes failure in this
specific code because when the task is run, shared P
might not be available, as it belongs to the thread.

program
implicit none
integer, target :: N(1) = (/1/)
integer, pointer :: P

!$omp parallel private(P) num_threads(4)
  P => N(1)
  !$omp task ! P should be firstprivate by default.
    P = 10
  !$omp end task
!$omp end parallel
print *, N
end program tfp