Fix issue with forward propagation through Cray pointer
When forward propagating an address through a Fortran
Cray pointer, Keep track of when an address value is
stored into an integer.

When replacing one ILI for another, if the original
was an address-valued ILI, and the new one is not,
then look for an IL_AKMV or IL_IKMV, or add an IL_KAMV
or IL_IAMV, as appropriate.