Cache cpuid calls
Experiments on Windows, at least, show this is a
performance boost for the driver.

Add X86IDFN(hw_features) that contains a bitmask of
supported processor features.

The global variable is initialized by calling:
X86IDFN(init_hw_features)(X86IDFN(hw_feature))

The math dispatch table assumed that if SSSE3 is
present all flavors of SSE ISA features are present.
This is a wrong assumption in that there are AMD
processors that have SSE4A but not SSSE3.

Break out SSSE3 in to a feature/category of its own.