Changes for LLVM 'vcvttsd2usi' emulation
LLVM emulates 'vcvttsd2usi' (a new AVX-512F
instruction) with 'vcvttsd2si' on non AVX-512F
machines to cast double to unsigned long. With
-Ktrap=fp option, this generates a floating
point exception when the converted number is >=
9223372036854775808 (1<<63). To get around this
problem, limit the maximum size of number to be
1<<63. Use the indirect way of converting 'double'
to 'uint64_t' (unsigned long) type in such cases.

Fix other places in the runtime where there was a
type cast from 'double' to 'uint64_t'.