Fix regression in pointer assignment to extended derived type
In newargs_for_entry(), make sure the descriptor
SDSC field for function results is updated with the
new descriptor dummy argument. This is only needed
for simple results such a pointers to scalars. In the
case with a pointer to a derived type, the descriptor
was created earlier (in semant1()).

Also make sure SDSC is properly updated because the
NEWDSC field is overloaded by the ADDRESS field. This
field is reset in lower_visit_symbol() in lowersym.c.