Better handle constants passed to unspecified intent dummy arguments
When passing constant to Fortran subroutine, assign
constant to a temp and then pass the address of temp as
argument when it's not pass-by-value.

The user code is not legal Fortran (dummy argument
is assigned ('defined'), but actual argment is a
constant/literal and is not defined. See Fortran
2018, 8.5.10, paragraph 5: "If no INTENT attribute is
specified for a dummy argument, its use is subject
to the limitations of its effective argument."),
we will move the constant from read-only memory,
and does not seg fault at run time.