Add debug support for Fortran allocatable array.
The bridge ulitizes the back link from array pointer symbol to the array
symbol to make a condition check if a Fortran array needs debug information.
Then when walking through the ilis, generate two variations of types for the
array. For the case with POINTER on declaration, emit type of pointer to a
FortranArrayType; for the case with ALLOCATABLE on declaration, emit a
FortranArrayType.

For type of FortranArrayType, the subranges (lower and upper bounds on each
dimension) are calculated based on the array descriptor. Struct F90_Desc and
Struct F90_DescDim are referenced, as well as the lower bound offset and the
extent offset, to compute the array bounds.

DW_OP_ operations are enriched to support the DW_OP list for computing and
dereference, such as DW_OP_minus and DW_OP_dup.