Fix infinite loop in final procedures
When generating layout descriptors, do not include type bound
procedures since they are not considered finalizable objects.
Otherwise, this can lead to problems in the runtime during
finalization. For example, if the type bound procedure is
associated with a function that returns the same type as its
derived type host, the runtime can go into an infinite loop during
finalization.