Fix ICE interf:new_symbol_and_link, symbol not found
Using a parameter array from a module which is constructed with implied-do
notation resulted in an ICE:

```
INTEGER*8, PARAMETER :: some_array(3) = (/ (i, i = 4241, 4243) /)
```

`const_eval()` computes the value of the index `i` in the implied do of the
array constructor. This function was returning a number literal, which was then
being fed into `mk_cval1()`.  This latter function expects a symbol table entry
when the number is a TY_INT8, and a number literal when it is a TY_INT.

This resulted in the number literal as it appeared in the source being
incorrectly treated as a symbol table index.

The fix is to use `mk_cval` which inserts the constant into the symbol table
and calls `mk_cval1` with what it expects.

This commit adds a test which sums an array which lives in another module and
checks that the sum is correct. `INTEGER` is used with `-fdefault-integer-8`,
which is how the issue was first reported.

Fixes https://github.com/flang-compiler/flang/issues/745

Signed-off-by: Peter Waller <peter.waller@arm.com>