Fix issue when binding name conflicts w/ public var
The issue is when a binding name binds to a type bound procedure
has the same name with another module public variable with a
parameter attribute which fails to export the module public
variable into a mod file. An error occurs when another module
is trying to use the public variable that cannot be found from
the mod file during the module import process.