Fix parsing of user defined operators
Fix parsing of user defined operators that are
defined in generic interfaces.

The <var ref> ::= <var primary> . <id> production
gets called twice for user defined operators,
such as x .MyOp. y. The compiler needs to keep
track of when this production is processed
in the beginning and end so it can properly
process the user defined operator. Otherwise,
it may issue a bogus semantic error.