Changes for handling nested parallel regions
In redundss (redundant subscript optimization), don't float
expressions out of a parallel region.

Keep track of the parallel region nest level, making FG_PAR
and FG_PARLOOP and LP_PARLOOP be integers instead of flags,
in the Fortran front end.

For most of the front end, they are tested for zero/nonzero,
so no changes needed in their use.