Fix incorrect rounding in ANINT()
This fix applies only to the "common", architecture
generic, version of anint/dnint.c

When the floating point argument to ANINT() is on the
range: [ 2^(number of mantissa bits) .. 2^(number
of mantissa bits +1) ) The adding of +/-0.5 to
the argument gets normalied in to the round bit
and depending on the processor's rounding mode can
incorrectly affect the input argument.

This fixes the ARM flang TA failures with test 'aint'.