Fix issue with LLVM bridge
In find_load_cse() the backward walk through the
previous instructions generated in the current basic
block is trying to look for the latest store into
the same memory location we are trying to load. But
the latest store may be through a CSExx, not just
a direct store. Look for that case.