Fix issue with bogus error message
Fix a bogus error message that is generated with
an argument check of an interface associated
with a procedure dummy argument in an internal
procedure.

The problem is that the interface symbol becomes
stale when the internal procedure is processed
between semantic pass 0 and 1. The compiler
will propagate symbol information from the
host procedure to the internal procedure but
not always from the internal procedure to the
host procedure. To fix this problem, defer the
argument check until processing the internal
procedure during semantic pass 1.