Fix issue with atan on x86
Fixed issue with the new atan function where
passing the argument(-0.0) was failing to
return -0.0. Fixed both the scalar and vector
versions.

Also changed which scalar function we call for
double precision atan when using the relaxed
version. It turns out that the previous version
of scalar double precision atan function was
slightly faster than our new one, therefore
decided to make a way that it could still
be called.