Ensure devcopy symbol is marked as internref if parent is
If an array is created using a directive, contained
subroutines are called which use this device
array. Turns out that every contained subroutine
gets its own copy of symbol $dev and thus any
initializations do not persist across subroutine
boundary. Therefore, at runtime, the program gets an
illegal address error because device pointer passed
to a kernel is not initialized.

This problem occurs because during upper phase,
when uplevel accesses are recorded. For an uplevel
symbol to be correctly created, it needs to have the
internref flag set. Turns out frontend sets this
flag, but it was omitting setting it for devcopy
symbol. Fix is to simply ensure this flag is also
set for devcopy symbol.