Remove quadratic loops from data-flow analysis
In add_use() each time a new use is created,
there is a check to see if the use has already
been added to prevent duplicates.  This linear
scan of uses caused quadratic execution time.
By replacing the linear scan with a hash table
lookup, we can remove the quadratic loop.

Change flow.c in the frontend and backend to
use the new use_hash in flow_utils.  Allocate
use_hash in flow_init(), and deallocate in
flow_end().  No longer use the linear scan of
the use list, but perform use_hash_insert() and
use_hash_lookup() calls instead.  The first_use
and num_uses variables are used to keep track
of new uses that are added for the linear scan.
Anywhere num_uses is incremented, we need a
use_hash_insert().  When first_use and num_uses
are reset, we need a use_hash_reset().  In a
couple functions, rename variable i to usex
to be more clear on what is happening within
the function.