Fixes for SIMD, especially on x86
Fix an issue where many SIMD intrinsics
were generating shufflevector with -1 as
parameters. Now generates an undef to match
Clang. A new flag is added to OperandFlag_t
to indicate if an OPERAND contains an undef,
so that undefs can be generated. A new struct
is added to the union inside of OPERAND, used
only when the undef flag is set. It stores the
operand's sptr and undef_mask, which indicates
which indices of the vector are undef.

All x86 intrinsics that use pointer types will
have that type converted to i8* to match llvm
signature.

x86_mmx type transform will also only happen
for x86 intrinsics only.