Fix false syntax error report with KIND_''
We incorrectly report an error when parsing
a KIND prefix before a zero-length string.

In crunch(), check for a possible kind value that
might precede an empty string literal. If there's
an ident_'' (i.e., an identifier followed with an
underscore in front of an empty string literal),
then generate a CH_UNDERSCORE pseudo-character
in front of the string literal. Prior to this
change, this functionality was already present
for strings with a len > 0; but not for empty
strings.