Fix issue with referencing type bound procedure
This error is exhibited when we reference
a type bound procedure within the arguments
to a call of the same type bound procedure, e.g.
    field_nodes = field%shape( size(field%shape()) )

In func_call2() and subr_call2(), change
put_memsym_of_ast() calls to a new function
called replace_memsym_of_ast().

The put_memsym_of_ast() routine replaces the
SPTR in an AST but this does not work if the
AST is shared. We are using this function to
replace a generic type bound procedure call to
its specific type bound procedure call. Using
put_memsym_of_ast() does not work in the case
where the AST is shared.

For example, we might have a generic type
bound procedure function call where one of
its arguments is the same generic type bound
procedure (e.g., x = generic_tbp(generic_tbp())).

To fix this bug, we use a new function called
replace_memsym_of_ast() that creates a duplicate
AST, but with a new symbol table pointer.

Add tests.