Fix issue with not resolving type bound procedures
There was an issue with resolving generic
type bound procedures passed as actual
arguments; subroutines were not correctly
matching procedure dummy arguments.

in tkr_match(), fix the check between the
dummy argument and the actual argument when
the dummy argument is an ST_PROC. If the
dummy argument and actual argument do not have
a DTYPE then they are both subroutines. So,
do not flag the dummy/actual argument passing
as an error if they both have a DTYPE of 0. We
were incorrectly assuming that the arguments
had to be functions. However, procedure dummy
arguments can take subroutines as well.