Expand OpenMP __tgt_target_data_update.
Implement ILI generation for
__tgt_target_data_update from BTARGETUPDATE,
TARGETUPDATE, and ETARGETUPDATE ILMs. The codegen
is enabled by -Mq,233,4 flag for now since the
runtime does not have this function yet. Once
it is implemented, this flag will be removed.

In addition, it changes the ILM definition of
TARGETUPDATE to make it the same design as
MP_MAP.