Invalid routine prolog code insertion (part 2)
The problem involves an automatic array of derived type that requires
inserting code in a routine prolog to 1) allocate memory for the array;
2) initialize a descriptor for the array; and 3) initialize array element
components.  The specific problem is that the descriptor is being
referenced in an initialization loop before it is set up.  The fix is
to change the order that these code insertions are made, by changing
when routine component_init_allocd_auto is called, relative to other
prolog code insertion cases.