Remove unnecessary temporaries for array sections
This was seen with a large application.

The function lhs_needtmp() calls
get_lhs_first_defs() to get a list of all the
reaching definitions to see if it is safe not
to generate a temp for array sections.  It hit
a FORALL statement and gave up, generating
a temp.  This fix added processing for such
FORALL statements.

Also added processing for pointer associate constructs.