Change for OpenMP target data in map clause
Check OpenMP's target data directive while
calculating array sections in map clause.

Array sections are still not supported in
Fortran, the whole array is sent to the
device, so semantic analysis for the array
section should be skipped for now. If not,
compilation fails as array sections are not
handled in the later phases in the compiler.
The right thing to do for the front end is to
skip it and gives a warning if the directive is
combined with "target" construct.

The compiler was not doing the right thing for
"target data", "target enter/exit data". This
patch adds check for them.

In short this patch fixes compilation fails for
the following codes:
    !$omp target enter data map(to:a(:,:) , b(:))