Fix issue where formatted write does not match item type
If the array descriptor comes from the parent
(host) subroutine, the INTERNREF flag of array
descriptor must be set. The missing case here
is when the array is a static member of derived
type variable and the derived type variable is
defined in parent routine, the array descriptor
is not set.  Handle this case.