Fix array init logic for module variable
Compilation of program with target offload can
lead to compile error with message:
Error: /tmp/xxxxx.gpu (12, 101): parse floating point constant invalid for type

The reason for this is because there is an array
in module that is being initialized. However,
when GPU llvm file is generated, only the
first element of that array is initialized with
that value.

The solution is to initialize the whole array
with same value if the constant value is also
array type. We also included an optimization for
ease of reading which uses "zeroinitializer"
if the value to initialize is zero - that way
individual elements do not need to printed out.