Mark start-of-loop with BIH_HEAD
This bit already existed in the flang codebase, but was unused.

Its documented intent is to mark the head of loops.
Match the behaviour to this documentation.

The approach is to find the branch targets of IM_DOEND and mark those.
This catches DO loops and various implicit loops.

This patch adds a test to check the ILI dump has the bit set. It is the first
test which looks at the ILI since it seems worth defending this behaviour.

A later commit makes use of this bit to add loop metadata to loop backedge
branches.