Remove assert in llutil for initializing array
In previous commit, updated array init logic
was developed.  An assert was added which
checked that the size of array being initialized
is the same size as that indicated by dinit
section. Turns out that this assert does not
always hold.

    equivalence (aux14_data(1,1),a1(1))

The same storage is used for two arrays. This
causes the frontend to break up array into
multiple logical chunks. In this case, the array
is broken up into one array of size 16384 and
another of size 1622016. The sum of these two
values corresponds to 16384 x 100.

Thus, the assert fails because the init thinks
the size is 1,638,400 but the array being
initialized is done in chunks and thus currently
the 1,622,016 sized chunk is being handled and
hence the size mismatch.

Simply remove the assert.