Fix internal compiler error on OpenPOWER
Fortran functions returning struct and having
a BIND(C) attribute should have return value
ILI as IL_GARGRET. This was not happening
when the function comes into expander with
"IM_CALL" opcode, because "IM_CALL" does not
have a case in the switch statement which
handles all the different call codes inside
the exp_call() routine.

On Power a function with BIND(C) attribute
returning a struct (with homogeneous members)
was failing because the return value was being
handled like an arg "IL_GARG" where the expected
type was an array (homogeneous aggregates are
returned as arrays on Power) and the actual
type struct, leading to a mismatch. Sending
the return value as "IL_GARGRET" will go down
a different path where the return value ILI is
handled correctly.